requirements.txt
# Core ML/AI
torch
transformers
opencv-python-headless
numpy
scikit-learn
open3d
Pillow
scipy

# Visualização Aprimorada
matplotlib

# Deploy e Demo
gradio
uvicorn
fastapi
python-dotenv
